Struct rusoto_elasticbeanstalk::ConfigurationSettingsDescription

pub struct ConfigurationSettingsDescription {
    pub application_name: Option<String>,
    pub date_created: Option<String>,
    pub date_updated: Option<String>,
    pub deployment_status: Option<String>,
    pub description: Option<String>,
    pub environment_name: Option<String>,
    pub option_settings: Option<Vec<ConfigurationOptionSetting>>,
    pub platform_arn: Option<String>,
    pub solution_stack_name: Option<String>,
    pub template_name: Option<String>,
}

Describes the settings for a configuration set.

Fields
application_name: Option<String>

The name of the application associated with this configuration set.

date_created: Option<String>

The date (in UTC time) when this configuration set was created.

date_updated: Option<String>

The date (in UTC time) when this configuration set was last modified.

deployment_status: Option<String>

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration that failed to successfully deploy.

description: Option<String>

Describes this configuration set.

environment_name: Option<String>

If not null, the name of the environment for this configuration set.

option_settings: Option<Vec<ConfigurationOptionSetting>>

A list of the configuration options and their values in this configuration set.

platform_arn: Option<String>

The ARN of the custom platform.

solution_stack_name: Option<String>

The name of the solution stack this configuration set uses.

template_name: Option<String>

If not null, the name of the configuration template for this configuration set.