Struct rusoto_elasticbeanstalk::CreateEnvironmentMessage

pub struct CreateEnvironmentMessage {
    pub application_name: String,
    pub cname_prefix: Option<String>,
    pub description: Option<String>,
    pub environment_name: Option<String>,
    pub group_name: Option<String>,
    pub option_settings: Option<Vec<ConfigurationOptionSetting>>,
    pub options_to_remove: Option<Vec<OptionSpecification>>,
    pub platform_arn: Option<String>,
    pub solution_stack_name: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub template_name: Option<String>,
    pub tier: Option<EnvironmentTier>,
    pub version_label: Option<String>,
}

Fields
application_name: String

The name of the application that contains the version to be deployed.

If no application is found with this name, CreateEnvironment returns an InvalidParameterValue error.

cname_prefix: Option<String>

If specified, the environment attempts to use this value as the prefix for the CNAME. If not specified, the CNAME is generated automatically by appending a random alphanumeric string to the environment name.

description: Option<String>

Describes this environment.

environment_name: Option<String>

A unique name for the deployment environment. Used in the application URL.

Constraint: Must be from 4 to 40 characters in length. The name can contain only letters, numbers, and hyphens. It cannot start or end with a hyphen. This name must be unique in your account. If the specified name already exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Default: If the CNAME parameter is not specified, the environment name becomes part of the CNAME, and therefore part of the visible URL for your application.

group_name: Option<String>

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name parameter. See Environment Manifest (env.yaml) for details.

option_settings: Option<Vec<ConfigurationOptionSetting>>

If specified, AWS Elastic Beanstalk sets the specified configuration options to the requested value in the configuration set for the new environment. These override the values obtained from the solution stack or the configuration template.

options_to_remove: Option<Vec<OptionSpecification>>

A list of custom user-defined configuration options to remove from the configuration set for this new environment.

platform_arn: Option<String>

The ARN of the custom platform.

solution_stack_name: Option<String>

This is an alternative to specifying a template name. If specified, AWS Elastic Beanstalk sets the configuration values to the default values associated with the specified solution stack.

tags: Option<Vec<Tag>>

This specifies the tags applied to resources in the environment.

template_name: Option<String>

The name of the configuration template to use in deployment. If no configuration template is found with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

tier: Option<EnvironmentTier>

This specifies the tier to use for creating this environment.

version_label: Option<String>

The name of the application version to deploy.

If the specified application has no associated application versions, AWS Elastic Beanstalk UpdateEnvironment returns an InvalidParameterValue error.

Default: If not specified, AWS Elastic Beanstalk attempts to launch the sample application in the container.