Struct rusoto_elasticbeanstalk::EnvironmentDescription

pub struct EnvironmentDescription {
    pub abortable_operation_in_progress: Option<bool>,
    pub application_name: Option<String>,
    pub cname: Option<String>,
    pub date_created: Option<String>,
    pub date_updated: Option<String>,
    pub description: Option<String>,
    pub endpoint_url: Option<String>,
    pub environment_id: Option<String>,
    pub environment_links: Option<Vec<EnvironmentLink>>,
    pub environment_name: Option<String>,
    pub health: Option<String>,
    pub health_status: Option<String>,
    pub platform_arn: Option<String>,
    pub resources: Option<EnvironmentResourcesDescription>,
    pub solution_stack_name: Option<String>,
    pub status: Option<String>,
    pub template_name: Option<String>,
    pub tier: Option<EnvironmentTier>,
    pub version_label: Option<String>,
}

Describes the properties of an environment.

Fields
abortable_operation_in_progress: Option<bool>

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel.

true: There is an update in progress.

false: There are no updates currently in progress.

application_name: Option<String>

The name of the application associated with this environment.

cname: Option<String>

The URL to the CNAME for this environment.

date_created: Option<String>

The creation date for this environment.

date_updated: Option<String>

The last modified date for this environment.

description: Option<String>

Describes this environment.

endpoint_url: Option<String>

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

environment_id: Option<String>

The ID of this environment.

environment_links: Option<Vec<EnvironmentLink>>

A list of links to other environments in the same group.

environment_name: Option<String>

The name of this environment.

health: Option<String>

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment.

Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment.

Green: Indicates the environment is healthy and fully functional.

Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request.

Default: Grey

health_status: Option<String>

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses.

platform_arn: Option<String>

The ARN of the custom platform.

resources: Option<EnvironmentResourcesDescription>

The description of the AWS resources used by this environment.

solution_stack_name: Option<String>

The name of the SolutionStack deployed with this environment.

status: Option<String>

The current operational status of the environment:

Launching: Environment is in the process of initial deployment.

Updating: Environment is in the process of updating its configuration settings or application version.

Ready: Environment is available to have an action performed on it, such as update or terminate.

Terminating: Environment is in the shut-down process.

Terminated: Environment is not running.

template_name: Option<String>

The name of the configuration template used to originally launch this environment.

tier: Option<EnvironmentTier>

Describes the current tier of this environment.

version_label: Option<String>

The application version deployed in this environment.