Struct rusoto_elasticbeanstalk::PlatformDescription

pub struct PlatformDescription {
    pub custom_ami_list: Option<Vec<CustomAmi>>,
    pub date_created: Option<String>,
    pub date_updated: Option<String>,
    pub description: Option<String>,
    pub frameworks: Option<Vec<PlatformFramework>>,
    pub maintainer: Option<String>,
    pub operating_system_name: Option<String>,
    pub operating_system_version: Option<String>,
    pub platform_arn: Option<String>,
    pub platform_category: Option<String>,
    pub platform_name: Option<String>,
    pub platform_owner: Option<String>,
    pub platform_status: Option<String>,
    pub platform_version: Option<String>,
    pub programming_languages: Option<Vec<PlatformProgrammingLanguage>>,
    pub solution_stack_name: Option<String>,
    pub supported_addon_list: Option<Vec<String>>,
    pub supported_tier_list: Option<Vec<String>>,
}

Detailed information about a platform.

Fields
custom_ami_list: Option<Vec<CustomAmi>>

The custom AMIs supported by the platform.

date_created: Option<String>

The date when the platform was created.

date_updated: Option<String>

The date when the platform was last updated.

description: Option<String>

The description of the platform.

frameworks: Option<Vec<PlatformFramework>>

The frameworks supported by the platform.

maintainer: Option<String>

Information about the maintainer of the platform.

operating_system_name: Option<String>

The operating system used by the platform.

operating_system_version: Option<String>

The version of the operating system used by the platform.

platform_arn: Option<String>

The ARN of the platform.

platform_category: Option<String>

The category of the platform.

platform_name: Option<String>

The name of the platform.

platform_owner: Option<String>

The AWS account ID of the person who created the platform.

platform_status: Option<String>

The status of the platform.

platform_version: Option<String>

The version of the platform.

programming_languages: Option<Vec<PlatformProgrammingLanguage>>

The programming languages supported by the platform.

solution_stack_name: Option<String>

The name of the solution stack used by the platform.

supported_addon_list: Option<Vec<String>>

The additions supported by the platform.

supported_tier_list: Option<Vec<String>>

The tiers supported by the platform.