Struct rusoto_elasticbeanstalk::UpdateEnvironmentMessage

pub struct UpdateEnvironmentMessage {
    pub application_name: Option<String>,
    pub description: Option<String>,
    pub environment_id: Option<String>,
    pub environment_name: Option<String>,
    pub group_name: Option<String>,
    pub option_settings: Option<Vec<ConfigurationOptionSetting>>,
    pub options_to_remove: Option<Vec<OptionSpecification>>,
    pub platform_arn: Option<String>,
    pub solution_stack_name: Option<String>,
    pub template_name: Option<String>,
    pub tier: Option<EnvironmentTier>,
    pub version_label: Option<String>,
}

Request to update an environment.

Fields
application_name: Option<String>

The name of the application with which the environment is associated.

description: Option<String>

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

environment_id: Option<String>

The ID of the environment to update.

If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

environment_name: Option<String>

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

group_name: Option<String>

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

option_settings: Option<Vec<ConfigurationOptionSetting>>

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

options_to_remove: Option<Vec<OptionSpecification>>

A list of custom user-defined configuration options to remove from the configuration set for this environment.

platform_arn: Option<String>

The ARN of the platform, if used.

solution_stack_name: Option<String>

This specifies the platform version that the environment will run after the environment is updated.

template_name: Option<String>

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

tier: Option<EnvironmentTier>

This specifies the tier to use to update the environment.

Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

version_label: Option<String>

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.