Crate rusoto_elasticbeanstalk

AWS Elastic Beanstalk

If you're using the service, you're probably looking for ElasticBeanstalkClient and ElasticBeanstalk.

Structs

AbortEnvironmentUpdateMessage

ApplicationDescription
Describes the properties of an application.

ApplicationDescriptionMessage
Result message containing a single description of an application.

ApplicationDescriptionsMessage
Result message containing a list of application descriptions.

ApplicationMetrics
Application request metrics for an AWS Elastic Beanstalk environment.

ApplicationResourceLifecycleConfig
The resource lifecycle configuration for an application. Defines lifecycle settings for resources that belong to the application, and the service role that Elastic Beanstalk assumes in order to apply lifecycle settings. The version lifecycle configuration defines lifecycle settings for application versions.

ApplicationResourceLifecycleDescriptionMessage

ApplicationVersionDescription
Describes the properties of an application version.

ApplicationVersionDescriptionMessage
Result message wrapping a single description of an application version.

ApplicationVersionDescriptionsMessage
Result message wrapping a list of application version descriptions.

ApplicationVersionLifecycleConfig
The application version lifecycle settings for an application. Defines the rules that Elastic Beanstalk applies to an application's versions in order to avoid hitting the per-region limit for application versions.

When Elastic Beanstalk deletes an application version from its database, you can no longer deploy that version to an environment. The source bundle remains in S3 unless you configure the rule to delete it.

ApplyEnvironmentManagedActionRequest
Request to execute a scheduled managed action immediately.

ApplyEnvironmentManagedActionResult
The result message containing information about the managed action.

AutoScalingGroup
Describes an Auto Scaling launch configuration.

BuildConfiguration
Settings for an AWS CodeBuild build.

Builder
The builder used to build the custom platform.

CPUUtilization
CPU utilization metrics for an instance.

CheckDNSAvailabilityMessage
Results message indicating whether a CNAME is available.

CheckDNSAvailabilityResultMessage
Indicates if the specified CNAME is available.

ComposeEnvironmentsMessage
Request to create or update a group of environments.

ConfigurationOptionDescription
Describes the possible values for a configuration option.

ConfigurationOptionSetting
A specification identifying an individual configuration option along with its current value. For a list of possible option values, go to Option Values in the AWS Elastic Beanstalk Developer Guide.

ConfigurationOptionsDescription
Describes the settings for a specified configuration set.

ConfigurationSettingsDescription
Describes the settings for a configuration set.

ConfigurationSettingsDescriptions
The results from a request to change the configuration settings of an environment.

ConfigurationSettingsValidationMessages
Provides a list of validation messages.

CreateApplicationMessage
Request to create an application.

CreateApplicationVersionMessage

CreateConfigurationTemplateMessage
Request to create a configuration template.

CreateEnvironmentMessage

CreatePlatformVersionRequest
Request to create a new platform version.

CreatePlatformVersionResult

CreateStorageLocationResultMessage
Results of a CreateStorageLocationResult call.

CustomAmi
A custom AMI available to platforms.

DeleteApplicationMessage
Request to delete an application.

DeleteApplicationVersionMessage
Request to delete an application version.

DeleteConfigurationTemplateMessage
Request to delete a configuration template.

DeleteEnvironmentConfigurationMessage
Request to delete a draft environment configuration.

DeletePlatformVersionRequest

DeletePlatformVersionResult

Deployment
Information about an application version deployment.

DescribeApplicationVersionsMessage
Request to describe application versions.

DescribeApplicationsMessage
Request to describe one or more applications.

DescribeConfigurationOptionsMessage
Result message containing a list of application version descriptions.

DescribeConfigurationSettingsMessage
Result message containing all of the configuration settings for a specified solution stack or configuration template.

DescribeEnvironmentHealthRequest
See the example below to learn how to create a request body.

DescribeEnvironmentHealthResult
Health details for an AWS Elastic Beanstalk environment.

DescribeEnvironmentManagedActionHistoryRequest
Request to list completed and failed managed actions.

DescribeEnvironmentManagedActionHistoryResult
A result message containing a list of completed and failed managed actions.

DescribeEnvironmentManagedActionsRequest
Request to list an environment's upcoming and in-progress managed actions.

DescribeEnvironmentManagedActionsResult
The result message containing a list of managed actions.

DescribeEnvironmentResourcesMessage
Request to describe the resources in an environment.

DescribeEnvironmentsMessage
Request to describe one or more environments.

DescribeEventsMessage
Request to retrieve a list of events for an environment.

DescribeInstancesHealthRequest
Parameters for a call to DescribeInstancesHealth.

DescribeInstancesHealthResult
Detailed health information about the Amazon EC2 instances in an AWS Elastic Beanstalk environment.

DescribePlatformVersionRequest

DescribePlatformVersionResult

ElasticBeanstalkClient
A client for the Elastic Beanstalk API.

EnvironmentDescription
Describes the properties of an environment.

EnvironmentDescriptionsMessage
Result message containing a list of environment descriptions.

EnvironmentInfoDescription
The information retrieved from the Amazon EC2 instances.

EnvironmentLink
A link to another environment, defined in the environment's manifest. Links provide connection information in system properties that can be used to connect to another environment in the same group. See Environment Manifest (env.yaml) for details.

EnvironmentResourceDescription
Describes the AWS resources in use by this environment. This data is live.

EnvironmentResourceDescriptionsMessage
Result message containing a list of environment resource descriptions.

EnvironmentResourcesDescription
Describes the AWS resources in use by this environment. This data is not live data.

EnvironmentTier
Describes the properties of an environment tier

EventDescription
Describes an event.

EventDescriptionsMessage
Result message wrapping a list of event descriptions.

Instance
The description of an Amazon EC2 instance.

InstanceHealthSummary
Represents summary information about the health of an instance. For more information, see Health Colors and Statuses.

Latency
Represents the average latency for the slowest X percent of requests over the last 10 seconds.

LaunchConfiguration
Describes an Auto Scaling launch configuration.

ListAvailableSolutionStacksResultMessage
A list of available AWS Elastic Beanstalk solution stacks.

ListPlatformVersionsRequest

ListPlatformVersionsResult

Listener
Describes the properties of a Listener for the LoadBalancer.

LoadBalancer
Describes a LoadBalancer.

LoadBalancerDescription
Describes the details of a LoadBalancer.

ManagedAction
The record of an upcoming or in-progress managed action.

ManagedActionHistoryItem
The record of a completed or failed managed action.

MaxAgeRule
A lifecycle rule that deletes application versions after the specified number of days.

MaxCountRule
A lifecycle rule that deletes the oldest application version when the maximum count is exceeded.

OptionRestrictionRegex
A regular expression representing a restriction on a string configuration option value.

OptionSpecification
A specification identifying an individual configuration option.

PlatformDescription
Detailed information about a platform.

PlatformFilter
Specify criteria to restrict the results when listing custom platforms.

The filter is evaluated as the expression:

Type Operator Values[i]

PlatformFramework
A framework supported by the custom platform.

PlatformProgrammingLanguage
A programming language supported by the platform.

PlatformSummary
Detailed information about a platform.

Queue
Describes a queue.

RebuildEnvironmentMessage

RequestEnvironmentInfoMessage
Request to retrieve logs from an environment and store them in your Elastic Beanstalk storage bucket.

RestartAppServerMessage

RetrieveEnvironmentInfoMessage
Request to download logs retrieved with RequestEnvironmentInfo.

RetrieveEnvironmentInfoResultMessage
Result message containing a description of the requested environment info.

S3Location
The bucket and key of an item stored in Amazon S3.

SingleInstanceHealth
Detailed health information about an Amazon EC2 instance in your Elastic Beanstalk environment.

SolutionStackDescription
Describes the solution stack.

SourceBuildInformation
Location of the source code for an application version.

SourceConfiguration
A specification for an environment configuration

StatusCodes
Represents the percentage of requests over the last 10 seconds that resulted in each type of status code response. For more information, see Status Code Definitions.

SwapEnvironmentCNAMEsMessage
Swaps the CNAMEs of two environments.

SystemStatus
CPU utilization and load average metrics for an Amazon EC2 instance.

Tag
Describes a tag applied to a resource in an environment.

TerminateEnvironmentMessage
Request to terminate an environment.

Trigger
Describes a trigger.

UpdateApplicationMessage
Request to update an application.

UpdateApplicationResourceLifecycleMessage

UpdateApplicationVersionMessage

UpdateConfigurationTemplateMessage
The result message containing the options for the specified solution stack.

UpdateEnvironmentMessage
Request to update an environment.

ValidateConfigurationSettingsMessage
A list of validation messages for a specified configuration template.

ValidationMessage
An error or warning for a desired configuration option value.

Enums

AbortEnvironmentUpdateError
Errors returned by AbortEnvironmentUpdate

ApplyEnvironmentManagedActionError
Errors returned by ApplyEnvironmentManagedAction

CheckDNSAvailabilityError
Errors returned by CheckDNSAvailability

ComposeEnvironmentsError
Errors returned by ComposeEnvironments

CreateApplicationError
Errors returned by CreateApplication

CreateApplicationVersionError
Errors returned by CreateApplicationVersion

CreateConfigurationTemplateError
Errors returned by CreateConfigurationTemplate

CreateEnvironmentError
Errors returned by CreateEnvironment

CreatePlatformVersionError
Errors returned by CreatePlatformVersion

CreateStorageLocationError
Errors returned by CreateStorageLocation

DeleteApplicationError
Errors returned by DeleteApplication

DeleteApplicationVersionError
Errors returned by DeleteApplicationVersion

DeleteConfigurationTemplateError
Errors returned by DeleteConfigurationTemplate

DeleteEnvironmentConfigurationError
Errors returned by DeleteEnvironmentConfiguration

DeletePlatformVersionError
Errors returned by DeletePlatformVersion

DescribeApplicationVersionsError
Errors returned by DescribeApplicationVersions

DescribeApplicationsError
Errors returned by DescribeApplications

DescribeConfigurationOptionsError
Errors returned by DescribeConfigurationOptions

DescribeConfigurationSettingsError
Errors returned by DescribeConfigurationSettings

DescribeEnvironmentHealthError
Errors returned by DescribeEnvironmentHealth

DescribeEnvironmentManagedActionHistoryError
Errors returned by DescribeEnvironmentManagedActionHistory

DescribeEnvironmentManagedActionsError
Errors returned by DescribeEnvironmentManagedActions

DescribeEnvironmentResourcesError
Errors returned by DescribeEnvironmentResources

DescribeEnvironmentsError
Errors returned by DescribeEnvironments

DescribeEventsError
Errors returned by DescribeEvents

DescribeInstancesHealthError
Errors returned by DescribeInstancesHealth

DescribePlatformVersionError
Errors returned by DescribePlatformVersion

ListAvailableSolutionStacksError
Errors returned by ListAvailableSolutionStacks

ListPlatformVersionsError
Errors returned by ListPlatformVersions

RebuildEnvironmentError
Errors returned by RebuildEnvironment

RequestEnvironmentInfoError
Errors returned by RequestEnvironmentInfo

RestartAppServerError
Errors returned by RestartAppServer

RetrieveEnvironmentInfoError
Errors returned by RetrieveEnvironmentInfo

SwapEnvironmentCNAMEsError
Errors returned by SwapEnvironmentCNAMEs

TerminateEnvironmentError
Errors returned by TerminateEnvironment

UpdateApplicationError
Errors returned by UpdateApplication

UpdateApplicationResourceLifecycleError
Errors returned by UpdateApplicationResourceLifecycle

UpdateApplicationVersionError
Errors returned by UpdateApplicationVersion

UpdateConfigurationTemplateError
Errors returned by UpdateConfigurationTemplate

UpdateEnvironmentError
Errors returned by UpdateEnvironment

ValidateConfigurationSettingsError
Errors returned by ValidateConfigurationSettings

Traits

ElasticBeanstalk
Trait representing the capabilities of the Elastic Beanstalk API. Elastic Beanstalk clients implement this trait.

Type Definitions

ARN

AbortableOperationInProgress

ActionHistoryStatus

ActionStatus

ActionType

ApplicationDescriptionList

ApplicationName

ApplicationNamesList

ApplicationVersionDescriptionList

ApplicationVersionProccess

ApplicationVersionStatus

AutoCreateApplication

AutoScalingGroupList

AvailableSolutionStackDetailsList

AvailableSolutionStackNamesList

BoxedBoolean

BoxedInt

Cause

Causes

CnameAvailability

ComputeType

ConfigurationDeploymentStatus

ConfigurationOptionDefaultValue

ConfigurationOptionDescriptionsList

ConfigurationOptionName

ConfigurationOptionPossibleValue

ConfigurationOptionPossibleValues

ConfigurationOptionSettingsList

ConfigurationOptionSeverity

ConfigurationOptionValue

ConfigurationOptionValueType

ConfigurationSettingsDescriptionList

ConfigurationTemplateName

ConfigurationTemplateNamesList

CreationDate

CustomAmiList

DNSCname

DNSCnamePrefix

DeleteSourceBundle

DeploymentTimestamp

Description

Ec2InstanceId

EndpointURL

EnvironmentDescriptionsList

EnvironmentHealth

EnvironmentHealthAttribute

EnvironmentHealthAttributes

EnvironmentHealthStatus

EnvironmentId

EnvironmentIdList

EnvironmentInfoDescriptionList

EnvironmentInfoType

EnvironmentLinks

EnvironmentName

EnvironmentNamesList

EnvironmentStatus

EventDate

EventDescriptionList

EventMessage

EventSeverity

ExceptionMessage

FailureType

FileTypeExtension

ForceTerminate

GroupName

ImageId

IncludeDeleted

IncludeDeletedBackTo

InstanceHealthList

InstanceId

InstanceList

InstancesHealthAttribute

InstancesHealthAttributes

Integer

LaunchConfigurationList

LaunchedAt

LoadAverage

LoadAverageValue

LoadBalancerList

LoadBalancerListenersDescription

Maintainer

ManagedActionHistoryItems

ManagedActions

MaxRecords

Message

NextToken

NonEmptyString

NullableDouble

NullableInteger

NullableLong

OperatingSystemName

OperatingSystemVersion

OptionNamespace

OptionRestrictionMaxLength

OptionRestrictionMaxValue

OptionRestrictionMinValue

OptionsSpecifierList

PlatformArn

PlatformCategory

PlatformFilterOperator

PlatformFilterType

PlatformFilterValue

PlatformFilterValueList

PlatformFilters

PlatformFrameworks

PlatformMaxRecords

PlatformName

PlatformOwner

PlatformProgrammingLanguages

PlatformStatus

PlatformSummaryList

PlatformVersion

QueueList

RefreshedAt

RegexLabel

RegexPattern

RequestCount

RequestId

ResourceId

ResourceName

S3Bucket

S3Key

SampleTimestamp

SolutionStackFileTypeList

SolutionStackName

SourceLocation

SourceRepository

SourceType

SupportedAddon

SupportedAddonList

SupportedTier

SupportedTierList

TagKey

TagValue

Tags

TerminateEnvForce

TerminateEnvironmentResources

TimeFilterEnd

TimeFilterStart

Timestamp

Token

TriggerList

UpdateDate

UserDefinedOption

ValidationMessageString

ValidationMessagesList

ValidationSeverity

VersionLabel

VersionLabels

VersionLabelsList

VirtualizationType