Struct rusoto_elasticbeanstalk::SingleInstanceHealth

pub struct SingleInstanceHealth {
    pub application_metrics: Option<ApplicationMetrics>,
    pub availability_zone: Option<String>,
    pub causes: Option<Vec<String>>,
    pub color: Option<String>,
    pub deployment: Option<Deployment>,
    pub health_status: Option<String>,
    pub instance_id: Option<String>,
    pub instance_type: Option<String>,
    pub launched_at: Option<String>,
    pub system: Option<SystemStatus>,
}

Detailed health information about an Amazon EC2 instance in your Elastic Beanstalk environment.

Fields
application_metrics: Option<ApplicationMetrics>

Request metrics from your application.

availability_zone: Option<String>

The availability zone in which the instance runs.

causes: Option<Vec<String>>

Represents the causes, which provide more information about the current health status.

color: Option<String>

Represents the color indicator that gives you information about the health of the EC2 instance. For more information, see Health Colors and Statuses.

deployment: Option<Deployment>

Information about the most recent deployment to an instance.

health_status: Option<String>

Returns the health status of the specified instance. For more information, see Health Colors and Statuses.

instance_id: Option<String>

The ID of the Amazon EC2 instance.

instance_type: Option<String>

The instance's type.

launched_at: Option<String>

The time at which the EC2 instance was launched.

system: Option<SystemStatus>

Operating system metrics from the instance.